giant rename resulting in a clean store.py that keeps all Chain and State (view) classes. also some doc fixes
